backend/requirements.txt
crewai==0.83.0
crewai-tools==0.12.1
langchain-groq==0.2.1
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn==0.32.1
supabase==2.10.0
pydantic==2.10.3
python-multipart==0.0.18
PyPDF2==3.0.1
python-docx==1.1.2
